% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_message.R
\name{debug_message}
\alias{debug_message}
\title{Title debug_message}
\usage{
debug_message(str.in, debug.level.trigger = 1)
}
\arguments{
\item{str.in}{- string input as}

\item{debug.level.trigger}{- integer}
}
\description{
Prints intended debug message

Whether to display the message is conditional on

debug.level.trigger <= DEBUG.LEVEL
}
\examples{
\dontrun{
my_fun <- function(some_tex_input){
   # Say "Hi!" at the beginning
   debug_message(paste0("my_fun(',some_tex_input,') started"))

   # Do some work with text
   debug_message("CP01: Checkpoint 1 passed", 5)

   # Do some work with text
   if(faulty condition detected) # Error detected here
   {
     debug_message("*** COND1 met ***: terminating", 2)

     # Say "Bye" at the end
     debug_message("my_fun() <- ERROR", -1)
   }

   debug_message("CP02: Checkpoint 2 passed", 5)
   # Do some work with text

   # Say "Bye" at the end
   debug_message("my_fun() <- OK", -1)
}
}
}
