% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_message.R
\name{debug_message_start}
\alias{debug_message_start}
\title{Title Message at the beginning of the debugged function}
\usage{
debug_message_start(...)
}
\description{
Title Message at the beginning of the debugged function
}
\examples{
 debug_message_start("Hi there!")
}
